refactor: Book でまとめた
